fix(frontend): keep relaCssLayer in one tsconfig project
CI's Frontend job failed on TS6305: the unit test imported relaCssLayer.ts
across the tsconfig project boundary, and vue-tsc -b requires a composite
project's declarations to be built first. It passed locally only because a
stale node_modules/.cache/tsc-node had them; a clean npm ci does not.

The file cannot live in both projects — including it in src/ gives TS6305,
excluding it gives TS6307 — so it stays in the node project alongside its test,
with vitest picking up root-level *.test.ts. The Vite plugin wrapper moves into
vite.config.ts, leaving relaCssLayer.ts as pure, testable logic.

Verified from a clean npm ci: typecheck, lint, 1575 tests, and build all pass.
